+++ packages/@dicebear/gridy/src/components/index.ts
@@ -1,0 +1,3 @@
+export { body } from './body';
+export { eyes } from './eyes';
+export { mouth } from './mouth';