+++ packages/@dicebear/gridy/src/index.ts
@@ -1,0 +1,18 @@
+/*!
+ * Gridy (@dicebear/gridy)
+ *
+ * Code licensed under MIT License.
+ * Copyright (c) 2021 Florian Körner
+ * Copyright (c) 2021 Franck Gabriel
+ *
+ * Design "Gridy" by Jan Forst licensed under MIT.
+ * Source: https://github.com/darosh/gridys/tree/master/packages/gridy-app-avatars
+ * License: https://github.com/darosh/gridys/blob/master/packages/gridy-app-avatars/LICENSE
+ */
+
+import { style } from './core';
+
+let { create, meta, schema } = style;
+
+export { create, meta, schema };
+export { Options } from './options';